Initial commit
Got a minor objectives actor set up aka, I made the actor and that's it. No functionality. 
The idea is have this control all the secondary objectives which can be turned on or off depending on the level designer's choice.  Dunno if it's worth trying to make it dynamically change as the level continues.